+++ src/options.ts
@@ -6,6 +6,7 @@ import { ShortcutsUI } from './options/ShortcutsUI';
 import { SessionsUI } from './options/SessionsUI';
 import { MigrateUI } from './options/MigrateUI';
 import { AboutUI } from './options/AboutUI';
+import { Theme } from './includes/Theme';
 
 async function init(): Promise<void>
 {
@@ -17,12 +18,27 @@ async function init(): Promise<void>
 	const config = await Configuration.load();
 	const sessions = await Sessions.load();
 	const local_files_allowed = await isLocalFilesAllowed();
+	const theme = await Theme.load();
 
-	new ConfigUI(document.getElementById('settings')!, config, local_files_allowed);
+	const config_ui = new ConfigUI(document.getElementById('settings')!, config, local_files_allowed, theme);
 	new ShortcutsUI(document.getElementById('shortcuts')!);
-	new SessionsUI(document.getElementById('sessions')!, sessions, config);
+	const session_ui = new SessionsUI(document.getElementById('sessions')!, sessions, config);
 	new MigrateUI(document.getElementById('migrate')!);
 	new AboutUI(document.getElementById('about')!);
+
+	chrome.storage.local.onChanged.addListener(async changes => {
+		if (Configuration.storageKey in changes)
+		{
+			const config = await Configuration.load();
+			config_ui.setConfig(config);
+			session_ui.setConfig(config);
+		}
+
+		if (Theme.storageKey in changes)
+		{
+			config_ui.setTheme(await Theme.load());
+		}
+	});
 
 	const migrate_comment_id = 'page_options_migrate_comment';
 	const migrate_comment = document.getElementById(migrate_comment_id);
@@ -36,6 +52,7 @@ async function init(): Promise<void>
 			.replace('__ID__', '<b>liekplgjlphohhlnfaibkdjindpnfimg</b>');
 	}
 
+	await Theme.apply(document.body);
 	i18n(document);
 }